requirements.txt
asyncpraw>=7.7.0
aiohttp>=3.8.0
feedparser>=6.0.0
jieba>=0.42.1
scikit-learn>=1.3.0
simhash>=2.1.0
openai>=1.0.0
anthropic>=0.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.9.0
pyyaml>=6.0.0
rich>=13.0.0
httpx>=0.25.0
